+++ riscv-rt/CHANGELOG.md
@@ -9,12 +9,12 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ### Added
 
-- Optional cargo feature `sbi` for SBI compatibility, including conditional compilation for machine mode instructions.
+- Optional cargo feature `s-mode` for supervisor mode, including conditional compilation for supervisor/machine mode instructions.
 
 ### Changed
 
 - Remove superfluous parentheses from link.x, which caused linker errors with nightly.
-- Changed `mp_hook` signature, hartid as passed as usize parameter by the caller (required for `sbi` feature).
+- Changed `mp_hook` signature, hartid as passed as usize parameter by the caller (required for `s-mode` feature).
 
 ## [v0.9.0] - 2022-07-01